Only show Coverage As for IResource objects
The Coverage As only makes sense for projects and files that can
be launched, but currently it is contributed to any Object type.
This includes viewers which aren't necessarily file based, or
have anything to do with projects that can be launched.

Modify the object type so that it only contributes to the action
when the selection is adaptable to IResource.

Fixes issue #87